Improve permission handling

+++ src/main/java/qupath/ext/omero/core/entities/repositoryentities/serverentities/ServerEntity.java
@@ -108,7 +108,7 @@ public abstract class ServerEntity implements RepositoryEntity {
     /**
      * @return the ID of the group owning this entity
      */
-    public int getGroupId() {
+    public long getGroupId() {
         return group.getId();
     }
